Fix warnings and errors for gcc-4.5 through gcc-4.7.
Fixes issue #465.

Patch contributions from Christian and dmp!

The build error was an assembler error in dynamo.c, caused by different versions
of gcc re-ordering the generated assembly.  My system gcc-4.3 switches to
.rodata to emit the string constants used in the following function before
switching back to .text for every function, so leaving the data section open is
no problem.  gcc-4.7, however, moves the inline asm to the top of the module and
emits very different assembly with no intermediate section switches.  The
resolution for now is just to switch to the .text section since that appears to
work with most major versions.  We can revisit this later if there is a problem.

The new warnings that were resolved:
-Wunused-but-set-variable
-Wenum-compare
-Waddress

-Wunused-but-set-variable is interesting, but it fires 145 times in DynamoRIO,
and the vast majority are for error codes that are checked with assertions or
logged in debug mode, but are ignored in release mode.  Therefore I disabled it.

-Wenum-compare is probably useful during development and was easy to fix in the
cases where we do want to compare different enums.

-Waddress caught an interesting NULL check of a function pointer, instead of a
similarly named variable.

SVN-Revision: 816

+++ core/x86/arch.c
@@ -464,7 +464,8 @@ arch_init()
 {
     ASSERT(sizeof(opnd_t) == EXPECTED_SIZEOF_OPND);
     /* ensure our flag sharing is done properly */
-    ASSERT(LINK_FINAL_INSTR_SHARED_FLAG < INSTR_FIRST_NON_LINK_SHARED_FLAG);
+    ASSERT((uint)LINK_FINAL_INSTR_SHARED_FLAG <
+           (uint)INSTR_FIRST_NON_LINK_SHARED_FLAG);
     ASSERT_TRUNCATE(byte, byte, OPSZ_LAST_ENUM);
     DODEBUG({ reg_check_reg_fixer(); });
